Extract GameService creation path from submit controller
