+++ convergence_games/apps/frontend/games/services/__init__.py
@@ -1,8 +1,11 @@
+from ._game_service import GameService, provide_game_service
 from ._search_service import SearchableBase, SearchResult, SearchService, provide_search_service
 
 __all__ = [
+    "GameService",
     "SearchResult",
     "SearchService",
     "SearchableBase",
+    "provide_game_service",
     "provide_search_service",
 ]
